build: add bl2/bl30 size check
PD# NONE

Problem:
When bl2 or bl30 is too big, fix_blx will fail, but mk script
will exit successfully

Solution:
add bl2.bin and bl30.bin size check in fix_blx

Verify:
Create a large bl2.bin and verify script exits with an error:
cp bl2/bin/g12a/bl2.bin ./
cat bl2/bin/g12a/bl2.bin >> bl2.bin
./mk g12a_skt_v1 --bl2 ./bl2.bin

+++ fip/gxl/build.sh
@@ -53,6 +53,11 @@ function fix_blx() {
 
 	# blx_size: blx.bin size, zero_size: fill with zeros
 	blx_size=`du -b $1 | awk '{print int($1)}'`
+	if [ $blx_size -gt $blx_bin_limit ]; then
+		echo "Error: $7 ($1) too big. $blx_size > $blx_bin_limit"
+		exit 1
+	fi
+
 	zero_size=$blx_bin_limit-$blx_size
 	dd if=/dev/zero of=$2 bs=1 count=$zero_size
 	cat $1 $2 > $3
